cloud_function_files/requirements.txt
beautifulsoup4==4.10.0
google-cloud-storage==2.2.1
lxml==4.8.0
requests==2.27.1